+++ src/controllers/ClasesController.js
@@ -1,26 +1,45 @@
-/*
+
 const Clases = require('../models/Clases');
 const Carrera = require('../models/Carrera');
 const Alumno = require('../models/Alumno');
 const UnidadCurricular = require('../models/UnidadCurricular');
 
-const getAll = ("" ,async(req, res) =>{
+const getAll = (async(req, res) =>{
     const result = await Clases.findAll()
     return res.json(result)
 })
 
 
 const getOne = async(req, res) =>{
-    const {id} = req.params
-    const result = await Clases.findByPk(id)
-    if(!result) return res.sendStatus(404)
-    return req.json(result)
+    try{
+        const {id} = req.params
+        if(!id){
+            return res.status(400).send('Debes proporcionar un id válido')
+        }
+        const result = await Clases.findByPk(id);
+        if(!result) {return res.sendStatus(404).send('No se encontró ninguna clase con ese id')
+        } else {
+            res.status(200).json(result);
+        }
+    }catch(error){
+        console.log(error)
+    }
 }
 
 const remove = async(req, res) =>{
-    const {id} = req.params
-    await Clases.destroy({where : {id}})
-    return res.sendStatus(204)
+    try{
+        const {id} = req.params
+        if(!id){
+            return res.status(400).send('Debes proporcionar un id válido')
+        }
+        const result = await Clases.destroy({where : {id}})
+        if(!result) {return res.sendStatus(404).send('No se encontró ninguna clase con ese id')
+        } else {
+            res.status(204);
+        }
+    }catch(error){
+        console.log(error)
+    }
 }
 
 module.exports = {
@@ -28,4 +47,4 @@ module.exports = {
     getOne,
     remove,
 }
-    */
+
