+++ src/controllers/UnidadCurricularController.js
@@ -11,18 +11,36 @@ const create = async(req, res) =>{
 }
 
 const getOne = async(req, res) =>{
-    const {id} = req.params
-    const result = await UnidadCurricular.findByPk(id)
-    if(!result) return res.sendStatus(404)
-    return req.json(result)
+    try{
+        const {id} = req.params
+        if(!id){
+            return res.status(400).send('Debes proporcionar un id válido')
+        }
+        const result = await UnidadCurricular.findByPk(id);
+        if(!result) {return res.sendStatus(404).send('No se encontró ninguna unidad curricular con ese id')
+        } else {
+            res.status(200).json(result);
+        }
+    }catch(error){
+        console.log(error)
+    }
 }
 
 const remove = async(req, res) =>{
-    const {id} = req.params
-    await UnidadCurricular.destroy({where : {id}})
-    return res.sendStatus(204)
+    try{
+        const {id} = req.params
+        if(!id){
+            return res.status(400).send('Debes proporcionar un id válido')
+        }
+        const result = await UnidadCurricular.destroy({where : {id}})
+        if(!result) {return res.sendStatus(404).send('No se encontró ninguna unidad curricular con ese id')
+        } else {
+            res.status(204);
+        }
+    }catch(error){
+        console.log(error)
+    }
 }
-
 const update = async(req, res) =>{
     const {id} = req.params
     const result = await UnidadCurricular.update(req.body, {where: {id}, returning: true})
